// src/sdk/schema.d.ts
export interface components {
    schemas: {
        /** Body_transcribe_transcribe_post */
        Body_transcribe_transcribe_post: {
            /**
             * File
             * Format: binary
             */
            file: string;
        };
        /** HTTPValidationError */
        HTTPValidationError: {
            /** Detail */
            detail?: components["schemas"]["ValidationError"][];
        };
        /** SynthesizeRequest */
        SynthesizeRequest: {
            /** Text */
            text: string;
            /**
             * Voice
             * @default af_bella
             */
            voice: string;
            /**
             * Speed
             * @default 1
             */
            speed: number;
        };
        /** ValidationError */
        ValidationError: {
            /** Location */
            loc: (string | number)[];
            /** Message */
            msg: string;
            /** Error Type */
            type: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}